requirements.txt
fastapi
uvicorn
httpx
pydantic
replicate
python-dotenv
